fix: rename RepositoryManager to Managers namespace, enhance GitHubRepository error handling, and clean up code formatting
